+++ src/api/app/Program.cs
@@ -3,6 +3,7 @@ using Intive.Patronage2023.Api.Configuration;
 using Intive.Patronage2023.Api.Keycloak;
 using Intive.Patronage2023.Api.User;
 using Intive.Patronage2023.Modules.Budget.Api;
+using Intive.Patronage2023.Modules.Budget.Application;
 using Intive.Patronage2023.Modules.Example.Api;
 using Intive.Patronage2023.Shared.Abstractions.Commands;
 using Intive.Patronage2023.Shared.Abstractions.Domain;
@@ -48,6 +49,7 @@ builder.Services.AddBudgetModule(builder.Configuration);
 builder.Services.Configure<ApiKeycloakSettings>(builder.Configuration.GetSection("Keycloak"));
 builder.Services.AddHttpClient();
 builder.Services.AddScoped<KeycloakService>();
+builder.Services.AddScoped<PermissionsService>();
 
 builder.Services.AddMediatR(cfg =>
 {
